renamed vepl2 to vepl
SVN-Revision: 29868

+++ contrib/tbl/vepl1/tests/test_driver.cxx
@@ -73,7 +73,7 @@ vil1_image CreateTestfloatImage(int wd, int ht)
   vil1_memory_image_of<float> image(wd, ht);
   for (int x = 0; x < wd; x++)
     for (int y = 0; y < ht; y++) {
-      float data = 0.01f * ((x-wd/2)*(y-ht/2)/16);
+      float data = 0.01f * float((x-wd/2)*(y-ht/2)/16);
       image.put_section(&data, x, y, 1, 1);
     }
   return image;
